Removing Administradores routes and promoting users to have access to the system

+++ app/Http/Controllers/Admin/UsuariosController.php
@@ -2,30 +2,36 @@
 
 namespace App\Http\Controllers\Admin;
 
-use App\Http\Controllers\Controller;
 use App\Http\Requests\PesquisasRequest;
 use App\Http\Requests\StoreUsuarioRequest;
 use App\Http\Requests\UpdateUsuarioRequest;
-use App\Repositories\StatusRepository;
 use App\Repositories\UsuarioRepository;
+use App\Repositories\GrupoRepository;
+use App\Repositories\StatusRepository;
+use App\Http\Controllers\Controller;
+use App\Services\ControleAcessoService;
+use Illuminate\Support\Facades\Auth;
 
 class UsuariosController extends Controller
 {
-    public $usuarios;
-    public $status;
+    protected $usuarios;
+    protected $grupos;
+    protected $status;
 
     /**
      * UsuariosController constructor.
      * @param UsuarioRepository $usuarioRepository
-     * @param StatusRepository $statusRepository
+     * @param GrupoRepository $grupoRepository
+     * @param StatusRepository $statusGeralRepository
      */
-    public function __construct(UsuarioRepository $usuarioRepository, StatusRepository $statusRepository)
+    public function __construct(UsuarioRepository $usuarioRepository, GrupoRepository $grupoRepository, StatusRepository $statusGeralRepository)
     {
         parent::__construct();
         $this->middleware('auth:admin');
 
         $this->usuarios = $usuarioRepository;
-        $this->status = $statusRepository;
+        $this->grupos = $grupoRepository;
+        $this->status = $statusGeralRepository;
     }
 
     /**
@@ -38,51 +44,67 @@ class UsuariosController extends Controller
         if ($parametro = $pesquisasRequest->search) {
             $usuarios->where('nome', 'like', "%{$parametro}%")->orWhere('sobrenome', 'like', "%{$parametro}%")->orWhere('email', 'like', "%{$parametro}%");
         }
-        return view('admin.usuarios.index', ['usuarios' => $usuarios->paginate(20)]);
+        return view('admin.usuarios.index', [
+            'usuarios' => $usuarios->paginate(20),
+        ]);
+
     }
 
     /**
+     * @param ControleAcessoService $controleAcessoService
      * @param StoreUsuarioRequest $storeUsuarioRequest
      * @return \Illuminate\Contracts\View\Factory|\Illuminate\Http\RedirectResponse|\Illuminate\View\View
      */
-    public function create(StoreUsuarioRequest $storeUsuarioRequest)
+    public function create(ControleAcessoService $controleAcessoService, StoreUsuarioRequest $storeUsuarioRequest)
     {
         if ($storeUsuarioRequest->isMethod('post')) {
-            if ($usuario = $this->usuarios->create($storeUsuarioRequest->only($this->usuarios->model()->getFillable()))) {
-                flash('Usuário criado com sucesso')->success();
+            if ($this->usuarios = $this->usuarios->model()->create($storeUsuarioRequest->only($this->usuarios->model()->getFillable()))) {
+                $this->usuarios->grupos()->sync($storeUsuarioRequest->grupos);
+                $this->usuarios->usuarios_permissoes()->sync(array_keys($storeUsuarioRequest->permissoes));
+                flash('Usuario criado com sucesso')->success();
                 return redirect()->route('admin.usuarios.index');
             }
-            flash('Não foi possível criar o usuário')->error();
+            flash('Não foi possível criar o usuario')->error();
             return redirect()->back();
         }
-        return view('admin.usuarios.create', ['status' => $this->status->model()->pluck('titulo', 'id')]);
+
+        return view('admin.usuarios.create', [
+            'status' => $this->status->model()->pluck('titulo', 'id'),
+            'grupos' => $this->grupos->model()->pluck('titulo', 'id'),
+            'permissoes' => $controleAcessoService->processaPermissoesArray()
+        ]);
     }
 
     /**
      * @param $id
      * @param UpdateUsuarioRequest $updateUsuarioRequest
+     * @param ControleAcessoService $controleAcessoService
      * @return \Illuminate\Contracts\View\Factory|\Illuminate\Http\RedirectResponse|\Illuminate\View\View
      */
-    public function edit($id, UpdateUsuarioRequest $updateUsuarioRequest)
+    public function edit($id, UpdateUsuarioRequest $updateUsuarioRequest, ControleAcessoService $controleAcessoService)
     {
-        $usuario = $this->usuarios->find($id);
         if ($updateUsuarioRequest->isMethod('post')) {
-            if ($usuario->find($id)->update($updateUsuarioRequest->only($this->usuarios->model()->getFillable()))) {
-                flash('Usuário atualizado com sucesso')->success();
-                return redirect()->route('admin.usuarios.index', [$id]);
+            if ($this->usuarios->model()->find($id)->update($updateUsuarioRequest->only($this->usuarios->model()->getFillable()))) {
+                $this->usuarios->model()->find($id)->grupos()->sync($updateUsuarioRequest->grupos);
+                $this->usuarios->model()->find($id)->usuarios_permissoes()->sync(array_keys($updateUsuarioRequest->permissoes));
+                flash('Usuario atualizado com sucesso')->success();
+                return redirect()->route('admin.usuarios.index');
             }
             flash('Não foi possível atualizar o usuário')->error();
             return redirect()->back();
         }
-        if ($usuario) {
+
+        if ($this->usuarios = $this->usuarios->model()->find($id)) {
             return view('admin.usuarios.edit', [
-                'usuario' => $usuario,
+                'grupos' => $this->grupos->model()->pluck('titulo', 'id'),
+                'grupo' => $this->usuarios->grupos()->first(),
+                'usuarios' => $this->usuarios,
                 'status' => $this->status->model()->pluck('titulo', 'id'),
+                'permissoes' => $controleAcessoService->processaPermissoesArray($id)
             ]);
         }
         return redirect()->route('admin.usuarios.index');
     }
-
 
     /**
      * @param $id
@@ -90,11 +112,16 @@ class UsuariosController extends Controller
      */
     public function destroy($id)
     {
+        if ($id == Auth::guard('admin')->user()->id) {
+            flash('Você não pode deletar seu próprio usuário.')->error();
+            return redirect()->back();
+        }
+
         if ($this->usuarios->model()->find($id)->delete()) {
-            flash('Usuário enviado para lixeira')->info();
+            flash('Usuario enviado para lixeira')->info();
             return redirect()->route('admin.usuarios.index');
         }
-        flash('Não foi possível remover este usuário')->error();
+        flash('Não foi possível remover este usuario')->error();
         return redirect()->back();
     }
 
@@ -104,10 +131,11 @@ class UsuariosController extends Controller
      */
     public function trashed(PesquisasRequest $pesquisasRequest)
     {
-        $usuarios = $this->usuarios->model()->onlyTrashed()->orderBy('deleted_at', 'DESC');
+        $usuarios = $this->usuarios->model()->orderBy('nome', 'ASC')->onlyTrashed()->orderBy('deleted_at', 'DESC');
         if ($parametro = $pesquisasRequest->search) {
             $usuarios->where('nome', 'like', "%{$parametro}%")->orWhere('sobrenome', 'like', "%{$parametro}%")->orWhere('email', 'like', "%{$parametro}%");
         }
+
         return view('admin.usuarios.lixeira', [
                 'usuarios' => $usuarios->paginate(20)
             ]
@@ -123,13 +151,13 @@ class UsuariosController extends Controller
         $usuarios = $this->usuarios->model()->onlyTrashed()->where('id', $id)->first();
         if (!empty($usuarios)) {
             if ($usuarios->restore()) {
-                flash('Usuário recuperado com sucesso.')->success();
+                flash('Usuario recuperado com sucesso.')->success();
                 return redirect()->route('admin.usuarios.trashed');
             }
-            flash('Não foi possível localizar o Usuário, parece que ele já foi removido')->error();
+            flash('Não foi possível localizar o Usuario, parece que ele já foi removido')->error();
             return redirect()->route('admin.usuarios.index');
         }
-        flash('Não foi possível recuperar o Usuário.')->error();
+        flash('Não foi possível recuperar o Usuario.')->error();
         return redirect()->route('admin.usuarios.index');
     }
 
@@ -140,10 +168,10 @@ class UsuariosController extends Controller
     public function forceDelete($id)
     {
         if ($this->usuarios->model()->onlyTrashed()->where('id', $id)->forceDelete()) {
-            flash('Usuário removido da lixeira com sucesso')->success();
+            flash('Usuario removido da lixeira com sucesso')->success();
             return redirect()->route('admin.usuarios.trashed');
         }
-        flash('Não é possível deletar este Usuário')->info()->important();
+        flash('Não é possível deletar este Usuario')->info()->important();
         return redirect()->back();
     }
 }
